refactor: some renaming and test fixes
